requirements.txt
# mkdir ~/.venv
# python3.7 -m venv ~/.venv/xiboside
# source ~/.venv/xiboside/bin/activate
# pip install -r requirements.txt

wheel==0.33.6
suds-py3==1.3.3.0
pycrypto==2.6.1
pyzmq==18.1.0
shiboken2==5.13.0
PySide2==5.13.0

